Seems to work
Considering the time I tried to develop a machine learning classifier, this is almost easier. You click through a file for a few minutes, but considering that you will spend days reading that book, the relative time investment is not that bad. The machine learning stuff was difficult to be 100% accurate, and that is not good when it means you will have missing text in your result.